requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pypdf2==3.0.1
transformers==4.44.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pinecone
langchain==0.0.308
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.8